net/hns3: add queue count of Rx API support
Here is to implement the available and used rxd number
count function. In Kunpeng series, the NIC hardware
support to read the bd numbers which wait processed from
the hardware FBD(Full Buffer Descriptor), and the driver
maintains the bd number to be wrriten back hardware.
compare the number of FBDs with the number of BDs
to be written back to the hardware.
the number of used descriptors of a rx queue is computed
as follows:
The fbd numbers of reading from FBD register plus the
bd numbers to be wrriten back to hardware maintainer by
the driver.

Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>