net/mlx5: fix switch port id when representor in bonding
In the bonding configurations the port switch id for representors
was composed of pf index in bonding as the 1 msb and the representor's
index as the remaining 15 lsbs. The special corner case for the host
PF representor on BF setups with representor id 0xFFFF was missed as well.

The new switch port id consists of 4 msbs for the pf bonding index and
the remaining 12 lsbs for the representor index. The switch port id ranges
for each type of representors are as follows:

Uplink representor(AKA master): 0xFFFF
Host PF representor: 0x<pf_bond>FFF
VF representor: 0x<pf_bond>[0-FFE]

Fixes: bee57a0a3565 ("net/mlx5: update switch port id in bonding configuration")
Cc: stable@dpdk.org

Signed-off-by: Bill Zhou <dongzhou@nvidia.com>
Acked-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>