app/testpmd: support age shared action context
When an age action becomes aged-out the rte_flow_get_aged_flows should
return the action context supplied by the configuration structure.

In case the age action created by the shared action API, the shared
action context of the Testpmd application was not set.

In addition, the application handler of the contexts returned by the
rte_flow_get_aged_flows API didn't consider the fact that the action
could be set by the shared action API and considered it as regular flow
context.

This caused a crash in Testpmd when the context is parsed.

This patch set context type in the flow and shared action context and
uses it to parse the aged-out contexts correctly.

Signed-off-by: Matan Azrad <matan@nvidia.com>
Acked-by: Dekel Peled <dekelp@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>