net/hns3: fix queue enabling state not store after FLR
FLR operation will reset the queue enabling state and
the driver needs to restore the state after reset. As
a result, it will result in unpredictable behavior with
reseted when user start or stop queue by calling the
relatived function if the driver does not do this.

This patch fix it by add a queue enabling state restore
function to the reset handler.

Fixes: fa29fe45a7b4 ("net/hns3: support queue start and stop")
Cc: stable@dpdk.org

Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Lijun Ou <oulijun@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>