Revert "net/mlx5: fix Rx queue count calculation"
This reverts commit d2d57605522d4a43be17e22e649e54033f6d8835.

This fix is uncorrect for at least two reasons.

First issue, when there are more than 8 CQEs to uncompress, the
computation done in this commit cannot work. Because the zip-ai
variable describes the current index inside the CQE8 array and thus is
limited from 0 to 7 included. So if we are decompressed the 9 packets,
ai is 0. So in this case, n is equals to cqe_cnt - 0.

Example with 11 packets we will have:
C | a | e0 | e1 | e2 | e3 | e4 | e5 | C | a | e0

c <-- CQE compressed
a <-- Array of minicqe
ex <-- emptry entry to store uncompressed CQE.

If the 9th packet is decompressed by the soft, n is equals to 9.
But with this commit, n is equals to 11 (i.e. 11 - 0).

Second issue is to count the next packet.

Example:

     packet 1                       |    packet 2
C | a | e0 | e1 | e2 | e3 | e4 | e5 | C | a | e0

There are 2 packets compressed in the first queue. For the first packet,
n is computed correctly.

But for the second, n is not computed properly. Because the zip context
is for the first packet. The  second packet is not yet decompressed, so
there are no context.

Signed-off-by: Maxime Leroy <maxime.leroy@6wind.com>
Signed-off-by: Nelio Laranjeiro <nelio.laranjeiro@6wind.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>