net/af_xdp: fix 32-bit build for older kernels
'uint64_t' is used to hold pointers in multiple locations in the
copy-mode code (used for kernels before 5.4). For a 32-bit build
this assumption is wrong and results in build errors. This commit
replaces such instances of 'uint64_t' with 'uintptr_t'.

While the copy-mode code will now compile for 32-bit, the PMD is
not expected to work and will fail at initialisation due to some
limitations in the kernel that were subsequently removed in v5.4.
Add a note to the docs to flag this limitation.

Fixes: f1debd77efaf ("net/af_xdp: introduce AF_XDP PMD")
Fixes: d8a210774e1d ("net/af_xdp: support unaligned umem chunks")
Cc: stable@dpdk.org

Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>