net/mlx5: fix validate RSS queues types
When the RSS queues' types are not uniformed, i.e, mixed with normal RX
queue and hairpin queue, PMD accept this flow after commit[1] instead of
reject it.

This because commit[1] create RX queue object as DEVX type via DEVX API
instead of IBV type via verbs, in which the latter will check the queues'
type when create verbs ind table but the formmer doesn't check when
create DEVX ind table.

However, in any case, logically PMD should check whether the input
configuration of RSS action is reasonable or not, which should
include queues' type check as well as the others.

So add the check of RSS queues' type in validation function to fix issue.

Fixes: 63bd16292c3a ("net/mlx5: support RSS on hairpin")
Cc: stable@dpdk.org

[1]:
commit 6deb19e1b2d2 ("net/mlx5: separate Rx queue object creations")

Signed-off-by: Xiaoyu Min <jackmin@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>