net/mlx5: fix sample and mirror use incorrect devices
Currently, sample and mirror actions are added to sh(shared context)
level list. The ports belong to the same sh can share the sample and
mirror actions.

While creating the sample and mirror actions sh list in sh creation,
the port which creates sh will be added to the sample and mirror list
context. The port device saved in the list context will be used when
release the entries in the list. Since mirror and sample actions may
have hrxq sub-action which is port level not sh level, in this case,
the other ports in the sh which creates the sample and mirror action
will use the incorrect port devices to release the hrxq.

This commit saves the create device in the sample and mirror actions
struct to avoid using the incorrect port device in entry releasing.

Fixes: 19784141692e ("net/mlx5: make sample and mirror action thread safe")

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Reviewed-by: Jiawei Wang <jiaweiw@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>